requirements.txt
PyMySQL==1.0.3
PyYAML==6.0
